# Chore: Improve utils-dom perf with matches() refactor