--- a/packages/utils-dom/src/cached-css-selectors.ts
+++ b/packages/utils-dom/src/cached-css-selectors.ts
@@ -1,4 +1,0 @@
-import { CompiledQuery } from 'css-select';
-
-// TODO: Use quick-lru so that it doesn't grow without bounds
-export const CACHED_CSS_SELECTORS: Map<string, CompiledQuery> = new Map();