--- a/packages/utils-dom/src/htmlelement.ts
+++ b/packages/utils-dom/src/htmlelement.ts
@@ -1,3 +1,4 @@
+import { compile } from 'css-select';
 import * as parse5 from 'parse5';
 import * as htmlparser2Adapter from 'parse5-htmlparser2-tree-adapter';
 
@@ -9,6 +10,7 @@ import { ElementData, INamedNodeMap } from './types';
 import { Node } from './node';
 import { CSSStyleDeclaration } from './cssstyledeclaration';
 import { HTMLDocument } from './htmldocument';
+import { CACHED_CSS_SELECTORS } from './cached-css-selectors';
 
 /**
  * https://developer.mozilla.org/docs/Web/API/HTMLElement
@@ -233,7 +235,11 @@ export class HTMLElement extends Node {
      * https://developer.mozilla.org/en-US/docs/Web/API/Element/matches
      */
     public matches(selector: string): boolean {
-        return this.ownerDocument.querySelectorAll(selector).includes(this);
+        if (!CACHED_CSS_SELECTORS.has(selector)) {
+            CACHED_CSS_SELECTORS.set(selector, compile(selector));
+        }
+
+        return CACHED_CSS_SELECTORS.get(selector)!(this._element);
     }
 
     /**