## Pull request checklist

Make sure you:

- [x] Signed the [Contributor License Agreement](https://cla.js.foundation/webhintio/hint)
- [x] Followed the [commit message guidelines](https://webhint.io/docs/contributor-guide/getting-started/pull-requests/#commit-messages)

For non-trivial changes, please make sure you also:

- [ ] Added/Updated related documentation.
- [x] Added/Updated related tests.

## Short description of the change(s)

This change improves the performance of `utils-dom` by refactoring `Element.matches`.

#### webhint.io
| before (ms) | after (ms) |
| ------------  | ---------- |
|    0.7199      |   0.7000   |

#### cnn.com
| before (ms) | after (ms) |
| ------------  | ---------- |
|    +60000    |  16.4900  |

_Previously cnn.com could take minutes to complete._

The problem was matches relied on querySelectorAll which will always check every dom node O(N). Now we compile the selector  test function from `css-select` directly and invoke it on the current element O(1). This performance is not bottlenecked by `compile`.

This change also pulls `CACHED_CSS_SELECTORS` into a shared module used by `htmldocument.ts` and `htmlelement.ts`,

Close #3904